fixes cancel, closes #3 #4 #6 #7

+++ myslurm.py
@@ -9,7 +9,6 @@ from tqdm import tqdm as pbar
 import matplotlib.pyplot as plt
 from os import path as op
 from collections import defaultdict, Counter
-from tqdm import tqdm as pbar
 from typing import Union
 
 import pythonimports as pyimp
@@ -17,23 +16,23 @@ import pythonimports as pyimp
 
 def get_seff(outs: list, desc=None):
     """From a list of .out files (ending in f'_{SLURM_JOB_ID}.out'), get seff output."""
-    infos = {}
+    seffs = {}
     for out in pbar(outs, desc=desc):
-        infos[out] = Seff(getpid(out))
-    return infos
+        seffs[out] = Seff(getpid(out))
+    return seffs
 
 
-def get_mems(infos: dict, units="MB", plot=True) -> list:
+def get_mems(seffs: dict, units="MB", plot=True) -> list:
     """From output by `get_seff()`, extract mem in `units` units; histogram if `plot` is True.
 
     Parameters
     ----------
-    - infos : dict of any key with values of class Seff
+    - seffs : dict of any key with values of class Seff
     - units : passed to Seff._convert_mem(). options: GB, MB, KB
 
     """
     mems = []
-    for key, info in infos.items():
+    for info in seffs.values():
         if "running" in info.state().lower() or "pending" in info.state().lower():
             continue
         mems.append(info.mem(units=units, per_core=False))
@@ -62,12 +61,12 @@ def clock_hrs(clock: str, unit="hrs") -> float:
     return hrs
 
 
-def get_times(infos: dict, unit="hrs", plot=True) -> list:
-    """From dict(infos) [val = seff output], get times in hours.
+def get_times(seffs: dict, unit="hrs", plot=True) -> list:
+    """From dict(seffs) [val = seff output], get times in hours.
 
     fix: add in other clock units"""
     times = []
-    for key, info in infos.items():
+    for key, info in seffs.items():
         if "running" in info.state().lower() or "pending" in info.state().lower():
             continue
         hrs = info.walltime(unit=unit)
@@ -102,12 +101,10 @@ def sbatch(shfiles: Union[str, list], sleep=0, printing=False) -> list:
         sbatched = False
         while sbatched is False:
             try:
-                pid = (
-                    subprocess.check_output([shutil.which("sbatch"), sh])
-                    .decode("utf-8")
-                    .replace("\n", "")
-                    .split()[-1]
-                )
+                pid = (subprocess.check_output([shutil.which("sbatch"), sh])
+                       .decode("utf-8")
+                       .replace("\n", "")
+                       .split()[-1])
                 sbatched = True
             except subprocess.CalledProcessError as e:
                 failcount += 1
@@ -121,9 +118,10 @@ def sbatch(shfiles: Union[str, list], sleep=0, printing=False) -> list:
                 jobs[q.job()].append(q.pid())
             if job in list(jobs.keys()):
                 sbatched = True
-                pid = jobs[job]
-                if len(jobs[job]) > 1:
-                    sq.cancel(match=pid)
+                pids = jobs[job]
+                if len(pids) > 1:
+                    # cancel all but the oldest job in the queue
+                    sq.cancel(grepping=sorted(pids)[:-1])
                 break
         if printing is True:
             print("sbatched %s" % sh)
@@ -131,31 +129,6 @@ def sbatch(shfiles: Union[str, list], sleep=0, printing=False) -> list:
         time.sleep(sleep)
         del pid
     return pids
-
-
-def getpids(user=os.environ["USER"]) -> list:
-    """From squeue -u $USER, return list of queue."""
-    pids = Squeue(user=user).pids()
-    if len(pids) != len(list(set(pids))):
-        print("len !- luni pids")
-    return pids
-
-
-def getjobs(user=os.environ["USER"]) -> list:
-    """From squeue -u $USER, return list of job names, alert if len != unique."""
-    jobs = Squeue(user=user).jobs()
-    if len(jobs) != len(list(set(jobs))):
-        print("len != luni jobs")
-    return jobs
-
-
-def qaccounts(pd=False, user=os.environ["USER"]) -> list:
-    """From squeue -u $USER, return list of billing accounts."""
-    if pd is False:
-        accounts = Squeue(user=user).accounts()
-    else:
-        accounts = Squeue(states=["PD"], user=user).accounts()
-    return accounts
 
 
 class Seff:
@@ -360,9 +333,7 @@ sqinfo = SQInfo  # backwards compatibility
 def adjustjob(acct, jobid):
     """Move job from one account to another."""
     acct = acct.replace("_cpu", "")
-    subprocess.Popen(
-        [shutil.which("scontrol"), "update", f"Account={acct}_cpu", f"JobId={jobid}"]
-    )
+    subprocess.Popen([shutil.which("scontrol"), "update", f"Account={acct}_cpu", f"JobId={jobid}"])
     pass
 
 
@@ -389,7 +360,6 @@ class Squeue:
 
     TODO
     ----
-    TODO: pass .cancel() and .update() a list of pids
     TODO: address Squeue.balance() TODOs
     TODO: update_job needs to handle skipping over jobs that started running or closed after class instantiation
     TODO: update_job needs to skip errors when eg trying to increase time of job beyond initial submission
@@ -555,11 +525,7 @@ class Squeue:
                 found += 1
                 pass
         if found != 10:
-            print(
-                pyimp.ColorText(
-                    "FAIL: Exceeded five subprocess.CalledProcessError errors."
-                ).fail.bold()
-            )
+            print(pyimp.ColorText("FAIL: Exceeded five subprocess.CalledProcessError errors.").fail.bold())
             return []
 
         sq = [s for s in sqout if s != ""]
@@ -613,10 +579,11 @@ class Squeue:
         return info
 
     @staticmethod
-    def _update_job(cmd, job, jobid):
+    def _update_job(cmd, job=None, jobid=None):
         """Execute 'scontrol update' or 'scancel' command for jobid."""
         # add jobid to command
-        cmd.append(f"jobid={jobid}" if "scancel" not in cmd else jobid)
+        if jobid is not None:
+            cmd.append(f"jobid={jobid}" if "scancel" not in cmd else jobid)
 
         failcount = 0
         # try to execute commands three times before giving up (slurm is a jerk sometimes)
@@ -631,16 +598,22 @@ class Squeue:
                 return True
             except subprocess.CalledProcessError:
                 # see if job is running or still in queue
-                jobq = Squeue(grepping=jobid)
-                if len(jobq) == 0:
-                    return "missing"
-                elif jobq[jobid].state() == "R":
-                    return "running", jobq[jobid]
-                # otherwise count as failure
-                failcount += 1
-        print(
-            pyimp.ColorText(f"FAIL: Update failed for cmd: {job} {jobid}").fail().bold()
-        )
+                if all([job is None, jobid is None, "scancel" in "".join(cmd)]):
+                    # for scancel -u $USER
+                    sq = Squeue()
+                    if len(sq) > 0:
+                        failcount += 1
+                    else:
+                        return True
+                else:
+                    jobq = Squeue(grepping=jobid)
+                    if len(jobq) == 0:
+                        return "missing"
+                    elif jobq[jobid].state() == "R":
+                        return "running"
+                    # otherwise count as failure
+                    failcount += 1
+        print(pyimp.ColorText(f"FAIL: Update failed for cmd: {job} {jobid}").fail().bold())
         return False
 
     @staticmethod
@@ -697,7 +670,7 @@ class Squeue:
             if isinstance(exclude, str):
                 exclude = [exclude]
             else:
-                assert isinstance(exclude, list), 'Squeue.balance() only expects `exclude` as `str` or `list`.'
+                assert isinstance(exclude, list), "Squeue.balance() only expects `exclude` as `str` or `list`."
 
         # set up grepping as list
         if grepping is None:
@@ -705,7 +678,7 @@ class Squeue:
         elif isinstance(grepping, str):
             grepping = [grepping]
         else:
-            assert isinstance(grepping, list), 'Squeue.balance() only expects `grepping` as `str` or `list`.'
+            assert isinstance(grepping, list), "Squeue.balance() only expects `grepping` as `str` or `list`."
 
         # add wanted accounts to `grepping`
         if onaccount is not None:
@@ -715,13 +688,13 @@ class Squeue:
             elif isinstance(onaccount, list):
                 grepping.extend(onaccount)
             else:
-                raise Exception('Squeue.balance() only expects `onaccount` as `str` or `list`.')
+                raise Exception("Squeue.balance() only expects `onaccount` as `str` or `list`.")
 
         # determine whether to keep/ignore priority status jobs
         if priority is True:
-            grepping.append('priority')
+            grepping.append("priority")
         elif priority is False:
-            exclude.append('priority')
+            exclude.append("priority")
 
         # limit queue to keyword match
         _sq = Squeue._grep_sq(sq.values(), grepping)
@@ -751,7 +724,7 @@ class Squeue:
         if user is False:  # scontrol commands
             cmd = cmd.split()
         elif desc == "scancel" and user is True:  # cancel all jobs
-            return Squeue._update_job(["scancel", "-u", os.environ["USER"]], None, None)
+            return Squeue._update_job(["scancel", "-u", os.environ["USER"]])
         # get subset of jobs returned from __init__()
         _sq = self._filter_jobs(self, **kwargs)
         # update each of the jobs
@@ -775,11 +748,7 @@ class Squeue:
                         updated_result is False
                     ), '`updated_result` must be one of {True, False, "missing", "running"}'
         else:
-            print(
-                pyimp.ColorText(
-                    "None of the jobs in Squeue class passed criteria."
-                ).warn()
-            )
+            print(pyimp.ColorText("None of the jobs in Squeue class passed criteria.").warn())
         pass
 
     @staticmethod
@@ -795,18 +764,6 @@ class Squeue:
 
         pass
 
-    def _save_default_accounts(save_dir=os.environ['HOME']):
-        """Among accounts available, choose which to use during balancing.
-        
-        The chosen accounts will be saved as op.join(save_dir, 'accounts.pkl'), and will be used
-            to balance accounts in the future when setting `parentdir` in Squeue.balance to `save_dir`.
-        """
-        import balance_queue as balq
-        
-        balq.get_avail_accounts(parentdir=save_dir, save=True)
-        
-        pass
-
     def keys(self):
         return list(self.sq.keys())
 
@@ -817,10 +774,10 @@ class Squeue:
         return self.sq.items()
 
     def running(self):
-        return self._filter_jobs(self, states='R')
+        return self._filter_jobs(self, states="R")
 
     def pending(self):
-        return self._filter_jobs(self, states='PD')
+        return self._filter_jobs(self, states="PD")
 
     def states(self, **kwargs):
         """Get a list of job states."""
@@ -882,7 +839,7 @@ class Squeue:
         Squeue._update_queue(self, cmd, "update", **kwargs)
         pass
 
-    def balance(self, parentdir=os.environ['HOME'], **kwargs):
+    def balance(self, parentdir=os.environ["HOME"], **kwargs):
         """Evenly distribute pending jobs across available slurm sbatch accounts.
 
         Parameters
@@ -929,7 +886,7 @@ class Squeue:
             if info.state() != "PD":
                 _sq.pop(pid)
         if len(_sq) == 0 or _sq is None:
-            print('\tno pending jobs in queue matching query')
+            print("\tno pending jobs in queue matching query")
             return
 
         # get accounts available for billing
@@ -937,11 +894,9 @@ class Squeue:
 
         if len(user_accts) > 1:
             # get per-account lists of jobs in pending status, return if all accounts have jobs (no need to balance)
-            accts, early_exit_decision = balq.getaccounts(
-                [q.info for q in _sq.values() if "pd" in q.state().lower()],
-                "",
-                user_accts,
-            )
+            accts, early_exit_decision = balq.getaccounts([q.info for q in _sq.values() if "pd" in q.state().lower()],
+                                                          "",
+                                                          user_accts)
             balq.announceacctlens(accts, early_exit_decision, priority=priority)
             if early_exit_decision is True:
                 return
@@ -953,9 +908,7 @@ class Squeue:
             balq.redistribute_jobs(accts, user_accts, balance)
 
             # announce final job counts
-            time.sleep(
-                2
-            )  # give system a little time to update (sometimes it can print original job counts)
+            time.sleep(2)  # give system a little time to update (sometimes it can print original job counts)
             if "priority" in kwargs:
                 # if the job no longer has same priority status, it won't be found in new queue query
                 kwargs.pop("priority")
@@ -963,23 +916,16 @@ class Squeue:
                 # if the job is no longer on the queried account, it won't be found in new queue query
                 kwargs.pop("onaccount")
             sq = self._filter_jobs(Squeue(), **kwargs)  # re-query the queue, filter
-            balq.announceacctlens(
-                *balq.getaccounts(
-                    [q.info for q in sq.values() if "pd" in q.state().lower()],
-                    "final",
-                    user_accts,
-                ),
-                priority=priority,
-            )  # print updated counts
+            balq.announceacctlens(*balq.getaccounts([q.info for q in sq.values() if "pd" in q.state().lower()],
+                                                    "final",
+                                                    user_accts),
+                                  priority=priority)  # print updated counts
             # update self
             for pid, q in _sq.items():
                 account = sq[pid].account()
                 self[q.pid()].info = Squeue._update_self(q.info, account=account)
         else:
-            print(
-                "\tthere is only one account (%s), no more accounts to balance queue."
-                % user_accts[0]
-            )
+            print(f"\tthere is only one account ({user_accts[0]}), no more accounts to balance queue.")
 
         pass
 
@@ -1000,34 +946,13 @@ class Squeue:
             account_counts[q.account()] += 1
 
         # print account stats
-        print(
-            pyimp.ColorText(
-                "There are %s accounts with jobs matching search criteria."
-                % len(stats.keys())
-            ).bold()
-        )
+        print(pyimp.ColorText("There are %s accounts with jobs matching search criteria." % len(stats.keys())).bold())
         for account, acct_stats in stats.items():
-            print(
-                "\t",
-                account,
-                "has",
-                account_counts[account],
-                "total jobs, which include:",
-            )
+            print("\t", account, "has", account_counts[account], "total jobs, which include:")
             for stat, count in acct_stats.items():
-                print(
-                    "\t\t",
-                    count,
-                    "jobs with",
-                    "%s" % "state =" if stat in ["R", "PD"] else "status =",
-                    stat,
-                )
+                print("\t\t", count, "jobs with", "%s" % "state =" if stat in ["R", "PD"] else "status =", stat)
         # print status counts
-        print(
-            pyimp.ColorText(
-                "\nIn total, there are %s jobs:" % sum(statuses.values())
-            ).bold()
-        )
+        print(pyimp.ColorText("\nIn total, there are %s jobs:" % sum(statuses.values())).bold())
         for status, count in statuses.items():
             print("\t", count, "jobs with status =", status)
         # print state counts
@@ -1041,9 +966,7 @@ class Squeue:
 
         for pid, q in pbar(_sq.items()):
             if "pd" in q.state().lower():  # only pending jobs can be held
-                updated_result = Squeue._update_job(
-                    [shutil.which("scontrol"), "hold"], q.job(), pid
-                )
+                updated_result = Squeue._update_job([shutil.which("scontrol"), "hold"], q.job(), pid)
                 if updated_result is True:
                     # update job info in Squeue container
                     self[q.pid()].info = Squeue._update_self(q.info, **kwargs)
@@ -1060,9 +983,7 @@ class Squeue:
         released = 0
         for pid, q in pbar(_sq.items()):
             if "held" in q.status().lower():  # JobHeldUser
-                updated_result = Squeue._update_job(
-                    [shutil.which("scontrol"), "release"], q.job(), pid
-                )
+                updated_result = Squeue._update_job([shutil.which("scontrol"), "release"], q.job(), pid)
                 if updated_result is True:
                     # update job info in Squeue container
                     self[q.pid()].info = Squeue._update_self(q.info, **kwargs)
